Refactor: clean up the Tornado code a bit
